backend/generator/pyproject.toml
[project]
name = "generator"
version = "0.1.0"
description = "Image and video generation server with AI models"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "boto3>=1.40.65",
    "diffusers>=0.35.2",
    "fastapi>=0.121.0",
    "nvidia-ml-py3>=7.352.0",
    "peft>=0.17.1",
    "pillow>=12.0.0",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "redis>=7.0.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]

[project.scripts]
server = "server:run"
